chore: update price field type to Decimal in cars resource files

+++ dev-demo/resources/cars_Ch.ts
@@ -58,7 +58,7 @@ export default {
       inputSuffix: 'USD',
       allowMinMaxQuery: true,
       editingNote: 'Price is in USD',
-      type: AdminForthDataTypes.FLOAT,
+      type: AdminForthDataTypes.DECIMAL,
       required: true,
     },
     {
